requirements.txt
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.8; python_version < '3.7'
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tinytag==1.5.0
urllib3==1.26.5
